alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
kombu==5.5.3
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
prompt_toolkit==3.0.51
pycparser==2.22
pymemcache==4.0.0
python-dateutil==2.9.0.post0
python-memcached==1.62
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
